# packages/api-contracts/openapi/antrein-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/businesses": {
      "get": {
        "operationId": "BusinessesController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Name search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "latitude",
            "required": false,
            "in": "query",
            "description": "Reserved in v1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "longitude",
            "required": false,
            "in": "query",
            "description": "Reserved in v1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "radiusKm",
            "required": false,
            "in": "query",
            "description": "Reserved in v1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "recommended",
                "rating_desc",
                "name_asc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List active businesses for discovery",
        "tags": [
          "businesses"
        ]
      },
      "post": {
        "operationId": "BusinessesController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a business with its primary outlet",
        "tags": [
          "businesses"
        ]
      }
    },
    "/api/v1/businesses/{businessId}": {
      "get": {
        "operationId": "BusinessesController_details",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get public business details",
        "tags": [
          "businesses"
        ]
      },
      "patch": {
        "operationId": "BusinessesController_update",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update business profile and policies",
        "tags": [
          "businesses"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/management": {
      "get": {
        "operationId": "BusinessesController_management",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the managed business representation",
        "tags": [
          "businesses"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/services": {
      "get": {
        "operationId": "ServicesController_list",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "outletId",
            "required": false,
            "in": "query",
            "description": "Reserved in v1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "staffId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activeOnly",
            "required": false,
            "in": "query",
            "schema": {
              "default": "true",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List services of a business",
        "tags": [
          "services"
        ]
      },
      "post": {
        "operationId": "ServicesController_create",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a service",
        "tags": [
          "services"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/services/{serviceId}": {
      "patch": {
        "operationId": "ServicesController_update",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a service",
        "tags": [
          "services"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/services/{serviceId}/deactivate": {
      "post": {
        "operationId": "ServicesController_deactivate",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate a service",
        "tags": [
          "services"
        ]
      }
    },
...
  },
  "info": {
    "title": "AntreIn API",
    "description": "Booking, payment, check-in, and real-time queue platform. Human-readable contract: docs/02-api-contract.md",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AddressDto": {
        "type": "object",
        "properties": {
          "formatted": {
            "type": "string",
            "example": "Jl. Example No. 10, Jakarta"
          },
          "latitude": {
            "type": "number",
            "example": -6.2
          },
          "longitude": {
            "type": "number",
            "example": 106.8
          }
        },
        "required": [
          "formatted"
        ]
      },
      "CreateOutletDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Main Outlet"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+622112345678"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          }
        },
        "required": [
          "name",
          "address"
        ]
      },
      "CreateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "AntreIn Barbershop"
          },
          "description": {
            "type": "string"
          },
          "logoFileId": {
            "type": "string",
            "example": "fil_01J..."
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Jakarta"
          },
          "primaryOutlet": {
            "$ref": "#/components/schemas/CreateOutletDto"
          }
        },
        "required": [
          "name",
          "primaryOutlet"
        ]
      },
      "BookingPolicyDto": {
        "type": "object",
        "properties": {
          "minimumLeadMinutes": {
            "type": "number",
            "example": 60
          },
          "maximumAdvanceDays": {
            "type": "number",
            "example": 30
          },
          "automaticConfirmation": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "DepositPolicyDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "example": true
          },
          "defaultType": {
            "type": "string",
            "enum": [
              "none",
              "fixed"
            ]
          },
          "defaultValue": {
            "type": "number",
            "example": 10000
          }
        },
        "required": [
          "enabled"
        ]
      },
      "CancellationPolicyDto": {
        "type": "object",
        "properties": {
          "fullRefundBeforeMinutes": {
            "type": "number",
            "example": 360
          },
          "partialRefundBeforeMinutes": {
            "type": "number",
            "example": 120
          },
          "partialRefundPercentage": {
            "type": "number",
            "example": 50
          },
          "noShowRefundPercentage": {
            "type": "number",
            "example": 0
          }
        }
      },
      "UpdateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logoFileId": {
            "type": "string"
          },
          "supportedPaymentOptions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "pay_at_location",
                "full_payment",
                "deposit"
              ]
            }
          },
          "bookingPolicy": {
            "$ref": "#/components/schemas/BookingPolicyDto"
          },
          "depositPolicy": {
            "$ref": "#/components/schemas/DepositPolicyDto"
          },
          "cancellationPolicy": {
            "$ref": "#/components/schemas/CancellationPolicyDto"
          }
        }
      },
...
      "MoneyDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 50000
          },
          "currency": {
            "type": "string",
            "example": "IDR"
          }
        },
        "required": [
          "amount",
          "currency"
        ]
      },
...
      "DepositDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "none",
              "fixed",
              "percentage"
            ]
          },
          "value": {
            "type": "number",
            "example": 10000
          }
        },
        "required": [
          "type",
          "value"
        ]
      },
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Haircut"
          },
          "description": {
            "type": "string"
          },
          "imageFileId": {
            "type": "string",
            "example": "fil_01J..."
          },
          "durationMinutes": {
            "type": "number",
            "example": 45
          },
          "price": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "deposit": {
            "$ref": "#/components/schemas/DepositDto"
          },
          "eligibleStaffIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "durationMinutes",
          "price"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageFileId": {
            "type": "string"
          },
          "durationMinutes": {
            "type": "number"
          },
          "price": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "deposit": {
            "$ref": "#/components/schemas/DepositDto"
          },
          "eligibleStaffIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}